+++ examples/server_stop.js
@@ -1,6 +1,6 @@
-var echo = require('../dist/index.js');
+const echo = require('../dist/index.js');
 
-var options = require('../laravel-echo-server');
+const options = require('../fixed-laravel-echo-server');
 
 echo.run(options).then(echo => {
   echo.stop();