+++ src/database/database.ts
@@ -10,32 +10,31 @@ export class Database implements DatabaseDriver {
   /**
    * Database driver.
    */
-  private driver: DatabaseDriver;
+  private _driver: DatabaseDriver;
 
   /**
    * Create a new database instance.
    */
   constructor(private options: any) {
-    if (options.database == 'redis') {
-      this.driver = new RedisDatabase(options);
-    } else if (options.database == 'sqlite') {
-      this.driver = new SQLiteDatabase(options);
-    } else {
+    if (options.database == 'redis')
+      this._driver = new RedisDatabase(options);
+    else if (options.database == 'sqlite')
+      this._driver = new SQLiteDatabase(options);
+    else
       Log.error('Database driver not set.');
-    }
   }
 
   /**
    * Get a value from the database.
    */
   get(key: string): Promise<any> {
-    return this.driver.get(key)
+    return this._driver.get(key)
   };
 
   /**
    * Set a value to the database.
    */
   set(key: string, value: any): void {
-    this.driver.set(key, value);
+    this._driver.set(key, value);
   };
 }